requirements.txt
google-cloud-aiplatform
requests
python-dotenv
fastapi
uvicorn
